# perf: WIP speedup sentry warmup ~10% further by deferring initial ua-parser import