@mattrobenolt 2020-05-07
So I don’t have a way to quantify this, but we need to be careful doing these kinds of optimizations. They’re not always good in practice. It might be worth some sort of detection if booting into shell, tests, etc, and handling it conditionally. 

The issue is that if we speed up boot time and defer all expensive computations until “later”, it can have a large impact for things like, handling web requests. Since it’ll be deferred until first few requests lazily causing degraded performance. For an HTTP server, I’d rather spend all the time upfront while booting up so when ready to accept requests, everything is as fast as possible.