@joshuarli 2020-05-07
> So I don’t have a way to quantify this, but we need to be careful doing these kinds of optimizations. They’re not always good in practice. It might be worth some sort of detection if booting into shell, tests, etc, and handling it conditionally.
> 
> The issue is that if we speed up boot time and defer all expensive computations until “later”, it can have a large impact for things like, handling web requests. Since it’ll be deferred until first few requests lazily causing degraded performance. For an HTTP server, I’d rather spend all the time upfront while booting up so when ready to accept requests, everything is as fast as possible.

Yeah, that's pretty much the main downside. I thought about it, and it would probably take a decent amount of effort to make it conditional... so this is definitely the laziest way.

One potential solution would be to warm up the import in getsentry's prod configuration. If that sounds worthwhile I can block this until I show that that works.