@mattrobenolt 2020-05-07
I don't really care or have any way to measure or judge this, just wanted to throw out the pros and cons so we're making informed decisions.