@joshuarli 2020-10-12
I'm still seeing a consistent ~300ms speedup. Reproduce by tossing an exit at the end of pytest_configure and rerunning like, `time pytest tests/sentry/utils/test_zip.py` a few times to get it in the CPU cache.